Rename attachedServersTreeItem.ts to match new functionality
